backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.1
pdf2image==1.16.3
reportlab==4.0.7
pandas==2.1.2
matplotlib==3.8.1 
PyJWT==2.10.0
bcrypt==3.2.2
